requirements.txt
# Web UI
streamlit>=1.30.0

# FastAPI Backend
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0

# LangChain RAG
langchain>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
langchain-core>=0.1.0

# Vector Database
chromadb>=0.4.22

# Embeddings
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.0.0

# Data
pandas>=2.0.0
numpy>=1.24.0

# API & Utils
requests>=2.31.0
python-dotenv>=1.0.0

# Optional AI
google-generativeai>=0.3.0
